// src/api/generated/model/mgmtSoftwareModuleMetadata.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * hawkBit REST APIs
 * Eclipse hawkBit™ is a domain-independent back-end framework for rolling out software updates to constrained edge devices as well as more powerful controllers and gateways connected to IP based networking infrastructure.

 * OpenAPI spec version: v1
 */

export interface MgmtSoftwareModuleMetadata {
  /** Metadata property key */
  key: string;
  /** Metadata property value */
  value?: string;
  /** Metadata property is visible to targets as part of software update action */
  targetVisible?: boolean;
}